Revert changes to Overview component.
