--- assets/js/modules/search-console/components/dashboard/SearchFunnelWidget/Overview/MaybeCreateGoalCTA.js
@@ -1,106 +1,0 @@
-/**
- * MaybeCreateGoalCTA component for SearchFunnelWidget.
- *
- * Site Kit by Google, Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-import PropTypes from 'prop-types';
-
-/**
- * Internal dependencies
- */
-import Data from 'googlesitekit-data';
-import CreateGoalCTA from '../CreateGoalCTA';
-import { Cell } from '../../../../../../material-components';
-import { CORE_MODULES } from '../../../../../../googlesitekit/modules/datastore/constants';
-import { CORE_USER } from '../../../../../../googlesitekit/datastore/user/constants';
-import useDashboardType, {
-	DASHBOARD_TYPE_MAIN,
-} from '../../../../../../hooks/useDashboardType';
-import useViewOnly from '../../../../../../hooks/useViewOnly';
-const { useSelect } = Data;
-
-export default function MaybeCreateGoalCTA( {
-	analyticsGoalsData,
-	error,
-	showRecoverableAnalytics,
-} ) {
-	const dashboardType = useDashboardType();
-
-	const viewOnly = useViewOnly();
-
-	const analyticsModuleAvailable = useSelect( ( select ) =>
-		select( CORE_MODULES ).isModuleAvailable( 'analytics' )
-	);
-
-	const analyticsModuleConnected = useSelect( ( select ) =>
-		select( CORE_MODULES ).isModuleConnected( 'analytics' )
-	);
-
-	const isAuthenticated = useSelect( ( select ) =>
-		select( CORE_USER ).isAuthenticated()
-	);
-
-	const canViewSharedAnalytics = useSelect( ( select ) => {
-		if ( ! analyticsModuleAvailable ) {
-			return false;
-		}
-
-		if ( ! viewOnly ) {
-			return true;
-		}
-
-		return select( CORE_USER ).canViewSharedModule( 'analytics' );
-	} );
-
-	const showAnalytics =
-		canViewSharedAnalytics &&
-		analyticsModuleConnected &&
-		! error &&
-		! showRecoverableAnalytics;
-
-	const showGoalsCTA =
-		isAuthenticated &&
-		showAnalytics &&
-		dashboardType === DASHBOARD_TYPE_MAIN &&
-		! analyticsGoalsData?.items?.length;
-
-	const quarterCellProps = {
-		smSize: 2,
-		mdSize: showGoalsCTA ? 4 : 2,
-		lgSize: 3,
-	};
-
-	if ( ! showAnalytics ) {
-		return null;
-	}
-	return (
-		<Cell { ...quarterCellProps } smSize={ 4 }>
-			{ showGoalsCTA && <CreateGoalCTA /> }
-		</Cell>
-	);
-}
-
-MaybeCreateGoalCTA.propTypes = {
-	analyticsGoalsData: PropTypes.oneOfType( [
-		PropTypes.arrayOf( PropTypes.object ),
-		PropTypes.object,
-	] ),
-	error: PropTypes.object,
-	showRecoverableAnalytics: PropTypes.bool,
-};
