+++ assets/js/modules/search-console/components/dashboard/SearchFunnelWidget/Overview/index.js
@@ -1,7 +1,7 @@
 /**
  * Overview component for SearchFunnelWidget.
  *
- * Site Kit by Google, Copyright 2023 Google LLC
+ * Site Kit by Google, Copyright 2021 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,23 +30,28 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
-import { Grid, Row, Cell } from '../../../../../../material-components';
-import { extractSearchConsoleDashboardData } from '../../../../util';
-import { calculateChange } from '../../../../../../util';
-import { CORE_MODULES } from '../../../../../../googlesitekit/modules/datastore/constants';
-import { MODULES_SEARCH_CONSOLE } from '../../../../datastore/constants';
+import { ProgressBar } from 'googlesitekit-components';
+import { Grid, Row, Cell } from '../../../../../material-components';
+import { extractSearchConsoleDashboardData } from '../../../util';
+import { calculateChange } from '../../../../../util';
+import { CORE_MODULES } from '../../../../../googlesitekit/modules/datastore/constants';
+import { MODULES_ANALYTICS } from '../../../../analytics/datastore/constants';
+import { CORE_LOCATION } from '../../../../../googlesitekit/datastore/location/constants';
+import { MODULES_SEARCH_CONSOLE } from '../../../datastore/constants';
 import useDashboardType, {
 	DASHBOARD_TYPE_MAIN,
 	DASHBOARD_TYPE_ENTITY,
-} from '../../../../../../hooks/useDashboardType';
-import DataBlock from '../../../../../../components/DataBlock';
-import useViewOnly from '../../../../../../hooks/useViewOnly';
-import { CORE_USER } from '../../../../../../googlesitekit/datastore/user/constants';
-import MaybeProgressBar from './MaybeProgressBar';
-import MaybeActivateAnalyticsCTA from './MaybeActivateAnalyticsCTA';
-import MaybeWidgetReportError from './MaybeWidgetReportError';
-import MaybeCreateGoalCTA from './MaybeCreateGoalCTA';
-import MaybeRecoverableModules from './MaybeRecoverableModules';
+} from '../../../../../hooks/useDashboardType';
+import ActivateAnalyticsCTA from './ActivateAnalyticsCTA';
+import CreateGoalCTA from './CreateGoalCTA';
+import DataBlock from '../../../../../components/DataBlock';
+import RecoverableModules from '../../../../../components/RecoverableModules';
+import {
+	BREAKPOINT_SMALL,
+	useBreakpoint,
+} from '../../../../../hooks/useBreakpoint';
+import useViewOnly from '../../../../../hooks/useViewOnly';
+import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
 const { useSelect, useInViewSelect } = Data;
 
 function getDatapointAndChange( [ report ], selectedStat, divider = 1 ) {
@@ -74,6 +79,7 @@ const Overview = ( {
 	showRecoverableAnalytics,
 } ) => {
 	const dashboardType = useDashboardType();
+	const breakpoint = useBreakpoint();
 
 	const viewOnly = useViewOnly();
 
@@ -99,7 +105,16 @@ const Overview = ( {
 	const analyticsModuleActive = useSelect( ( select ) =>
 		select( CORE_MODULES ).isModuleActive( 'analytics' )
 	);
+	const analyticsModuleActiveAndConnected =
+		analyticsModuleActive && analyticsModuleConnected;
 
+	const isNavigatingToReauthURL = useSelect( ( select ) => {
+		if ( ! analyticsModuleAvailable ) {
+			return false;
+		}
+		const adminReauthURL = select( MODULES_ANALYTICS ).getAdminReauthURL();
+		return select( CORE_LOCATION ).isNavigatingTo( adminReauthURL );
+	} );
 	const isSearchConsoleGatheringData = useInViewSelect( ( select ) =>
 		select( MODULES_SEARCH_CONSOLE ).isGatheringData()
 	);
@@ -302,15 +317,52 @@ const Overview = ( {
 					</Row>
 				</Cell>
 
-				<MaybeProgressBar />
-				<MaybeActivateAnalyticsCTA />
-				<MaybeWidgetReportError
-					WidgetReportError={ WidgetReportError }
-					error={ error }
-					showRecoverableAnalytics={ showRecoverableAnalytics }
-				/>
-				<MaybeCreateGoalCTA />
-				<MaybeRecoverableModules />
+				{ isNavigatingToReauthURL && (
+					<Cell
+						{ ...halfCellProps }
+						className="googlesitekit-data-block__loading"
+					>
+						<ProgressBar />
+					</Cell>
+				) }
+
+				{ canViewSharedAnalytics &&
+					( ! analyticsModuleConnected || ! analyticsModuleActive ) &&
+					! isNavigatingToReauthURL && (
+						<Cell { ...halfCellProps }>
+							{ BREAKPOINT_SMALL !== breakpoint && (
+								<ActivateAnalyticsCTA />
+							) }
+						</Cell>
+					) }
+
+				{ ! showRecoverableAnalytics &&
+					canViewSharedAnalytics &&
+					analyticsModuleActiveAndConnected &&
+					error && (
+						<Cell { ...halfCellProps }>
+							<WidgetReportError
+								moduleSlug="analytics"
+								error={ error }
+							/>
+						</Cell>
+					) }
+
+				{ showAnalytics && (
+					<Cell { ...quarterCellProps } smSize={ 4 }>
+						{ showGoalsCTA && <CreateGoalCTA /> }
+					</Cell>
+				) }
+
+				{ canViewSharedAnalytics &&
+					analyticsModuleActiveAndConnected &&
+					showRecoverableAnalytics && (
+						<Cell { ...halfCellProps }>
+							<RecoverableModules
+								moduleSlugs={ [ 'analytics' ] }
+							/>
+						</Cell>
+					) }
 			</Row>
 		</Grid>
 	);
